server/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==23.5
pillow==10.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
